requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
SQLAlchemy==1.4.41
flask-sqlalchemy
sqlalchemy-utils
Flask-Migrate==2.6.0
Flask-Script
flask-cors
pyjwt
Pillow